update coordinates to floats for precise placement

+++ src/main/java/org/sk/skMinecraft/commands/EditEntity.java
@@ -16,7 +16,7 @@ public class EditEntity extends Command {
     private boolean ai;
 
     private boolean setPosition;
-    private int x,y,z;
+    private double x,y,z;
     private World world;
 
     private String name;
